# sdks/openapi.yaml
openapi: "3.0.3"
info:
  title: LLMSafeSpaces API
  description: |
    LLMSafeSpaces is a Kubernetes-first platform for running AI agents securely in isolated sandboxes.
    Every sandbox runs `opencode serve` as a persistent HTTP server with a PVC-backed persistent workspace.
  version: "0.7.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /workspaces/{id}:
    get:
      tags: [workspaces]
      summary: Get a workspace
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [workspaces]
      summary: Rename a workspace
      operationId: renameWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "204":
          description: Workspace renamed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [workspaces]
      summary: Delete a workspace
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "204":
          description: Workspace deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /workspaces/{id}/models:
    get:
      tags: [workspaces]
      summary: List available models for a workspace
      operationId: listWorkspaceModels
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /workspaces/{id}/model:
    put:
      tags: [workspaces]
      summary: Set the workspace's default model
      operationId: setWorkspaceModel
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model]
              properties:
                model:
                  type: string
                  description: Model ID (e.g. "claude-sonnet-4-20250514")
      responses:
        "204":
          description: Model set
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /provider-credentials/{id}/models:
    get:
      tags: [credentials]
      summary: Probe models available with a stored credential
      operationId: probeProviderCredentialModels
      x-feature-gated: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProbeModelsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/provider-credentials/{id}/models:
    get:
      tags: [credentials]
      summary: Probe models for an admin credential
      operationId: probeAdminProviderCredentialModels
      x-feature-gated: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProbeModelsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /orgs/{id}/credentials/{credID}/models:
    get:
      tags: [credentials]
      summary: Probe models reachable via an org credential (org admin)
      operationId: probeOrgCredentialModels
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - name: credID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Probe result (200 with `warning` on probe failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProbeModelsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    WorkspaceId:
      name: id
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
...
    OrgId:
      name: id
      in: path
      required: true
      description: Organization ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login or /auth/register
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "lsp_ prefixed API key in Bearer format (e.g. Bearer lsp_abc123...)"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        userId:
          type: string
        runtime:
          type: string
        storageSize:
          type: string
        phase:
          type: string
          enum: [Pending, Creating, Active, Suspending, Suspended, Resuming, Terminating, Terminated, Failed]
        pvcName:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        devPreviewEnabled:
          type: boolean
          description: Whether the authenticated HTTP/WS dev-preview tunnel is enabled (Epic 66).
          default: false
        defaultModel:
          type: string
        agentNeedsRefresh:
          type: boolean
          description: Agent credential state (Epic 27a); always present (no omitempty).
        credentialsPendingSince:
          type: string
          format: date-time
          nullable: true
...
    ModelListResponse:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelItem"
        currentModel:
          type: string
    ModelItem:
      type: object
      properties:
        id:
          type: string
        providerID:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        tier:
          type: string
        freeTier:
          type: boolean
        selected:
          type: boolean
...
    ProbeModelsResponse:
      type: object
      properties:
        models:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
...